Add files via upload
For this assignment, use the same bank marketing dataset, that you encountered in the Acquisition Analytics module.
You wanted to predict the probability of response of each prospect and target the ones most likely to respond to the next telemarketing campaign. The steps were as follows:

Identify the relevant predictor variables for a response using EDA
Build predictive models and choose the best one
Sort the prospects in decreasing probability of response (predicted by the best model) and target the top X% (or top Y deciles), where X would be determined by your business objective (e.g. maximizing the overall response rate/the number of responders at a fixed marketing cost)